requirements.txt
aiohttp==3.7.4.post0
alabaster==0.7.12
amqp==5.0.6
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.1
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
DAWG-Python==0.7.2
defusedxml==0.7.1
Deprecated==1.2.12
distlib==0.3.1
distro==1.5.0
Django==3.1.8
django-allauth==0.44.0
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-mptt==0.12.0
django-resized==0.3.11
docopt==0.6.2
docutils==0.16
filelock==3.0.12
flake8==3.9.1
gitdb==4.0.7
GitPython==3.1.17
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.2.52
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.8.1
PyJWT==2.0.1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==2.4.7
pytesseract==0.3.7
pytest==6.2.3
pytest-django==4.2.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-memcached==1.59
python-telegram-bot==13.5
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
smmap==4.0.0
snowballstemmer==2.1.0
Sphinx==3.5.4
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
telepot==12.7
toml==0.10.2
tornado==6.1
tox==3.23.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
vine==5.0.0
virtualenv==20.4.3
wcwidth==0.2.5
wrapt==1.12.1
yarl==1.6.3
yookassa==2.0.1
